Some updates to the Effects API to make it slightly easier to use in code (and thus slightly easier to deal with externally): startEffect() is no longer callable directly. Use begin() instead. EffectCreatorParameters is no longer needed by user code. Changed the way parameters are handled. The Effect class can add parameters to individual EFCs.
Removed Vectors in some places, replaced with the proper ArrayList. This will be ongoing as part of a need to massively overhaul some of the old code.

Started beginning of a new command system in which each command is a separate class. The core is already mostly complete, with a new lightweight CommandHandler and the ability to index Command objects from packages. Next step is to split all the old methods into their own classes. This new code is found under ring.commands.nc and is not integrated anywhere else yet.