New world builder DSL, shimmed into old with temp test world.
Create a new DSL to allow easy building of worlds in Kotlin script.
Currently, the loading of the world from eXist is now replaced with a
direct call to building a test world, and that test world is converted
to the old domain objects for compatibility.

The intention is to slowly refactor the old code until it converges
with the new types used by the DSL, at which point everything
collapses together into one.